# Reuse weak references in the WeakList