I guess it is better to retarget existing `WeakReference` objects whenever possible.