@smoogipoo 2018-10-18
Have you profiled that this is actually a problem? I think you're rather making the `Add()` method quadratic because of that extra iteration over every element of the list.