@UselessToucan 2018-10-19
Not sure that I benchmarked this correctly, but I will post the result anyway
Git repo https://github.com/UselessToucan/WeakListBenchmark
Benchmark code:
```
        public void BenchmarkMethod()
        {
            var sut = new WeakList<string>();
            //var sut = new TocoToucanWeakList<string>();

            var strongRefs = new List<string>(250);

            for (var i = 0; i < 500; i++)
            {
                var str = $"Before GC collection {i}";
                sut.Add(str);
                if (i % 2 == 0)
                    strongRefs.Add(str);
            }

            GC.Collect();

            sut.ForEachAlive(s => { });

            for (var i = 0; i < 250; i++)
            {
                var str = $"After GC collection {i}";
                sut.Add(str);
            }
        }
```

Results:
master
``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.345 (1803/April2018Update/Redstone4)
Intel Core i5-4200U CPU 1.60GHz (Haswell), 1 CPU, 4 logical and 2 physical cores
Frequency=2240906 Hz, Resolution=446.2481 ns, Timer=TSC
.NET Core SDK=2.1.403
  [Host]     : .NET Core ? (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.5 (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT


```
|          Method |     Mean |    Error |   StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |    Op/s |
|---------------- |---------:|---------:|---------:|----------:|---------:|---------:|---------:|---------:|---------:|--------:|
| BenchmarkMethod | 723.5 us | 3.098 us | 2.746 us | 0.7340 us | 720.5 us | 721.6 us | 722.7 us | 724.0 us | 728.9 us | 1,382.1 |


pr
``` ini

BenchmarkDotNet=v0.11.1, OS=Windows 10.0.17134.345 (1803/April2018Update/Redstone4)
Intel Core i5-4200U CPU 1.60GHz (Haswell), 1 CPU, 4 logical and 2 physical cores
Frequency=2240906 Hz, Resolution=446.2481 ns, Timer=TSC
.NET Core SDK=2.1.403
  [Host]     : .NET Core ? (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.5 (CoreCLR 4.6.26919.02, CoreFX 4.6.26919.02), 64bit RyuJIT


```
|          Method |     Mean |     Error |    StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |  Op/s |
|---------------- |---------:|----------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|------:|
| BenchmarkMethod | 3.353 ms | 0.0066 ms | 0.0058 ms | 0.0016 ms | 3.343 ms | 3.348 ms | 3.351 ms | 3.356 ms | 3.363 ms | 298.3 |

The idea was not to allocate/garbage collect `WeakReference` objects whenever possible, but it looks like it does not not worth it.